refactor: move provider to separate folder (part II)
